:recycle: #60 PicTopBarIcon model 패키지로 분리
